requirements.txt
# LeetCode Benchmark Tool Dependencies

# HuggingFace datasets for LeetCodeDataset
datasets>=2.14.0

# DeepSeek API (OpenAI-compatible)
openai>=1.0.0

# Progress bars
tqdm>=4.65.0

# Result formatting
tabulate>=0.9.0

# Environment variables
python-dotenv>=1.0.0
